requirements.txt
fastapi
uvicorn
faiss-cpu
pinecone-client
langchain
llama-index
openai
python-dotenv
tiktoken
pydantic
pytest
